Refactor FollowUp into a concern and implement interaction callbacks
